Move make_unique to source/util.
This MakeUnique code is used in places other then source/opt so move it
to source/utils.
