+++ source/opt/basic_block.cpp
@@ -18,9 +18,9 @@
 
 #include "source/opt/function.h"
 #include "source/opt/ir_context.h"
-#include "source/opt/make_unique.h"
 #include "source/opt/module.h"
 #include "source/opt/reflect.h"
+#include "source/util/make_unique.h"
 
 namespace spvtools {
 namespace opt {
