+++ source/opt/fold_spec_constant_op_and_composite_pass.cpp
@@ -21,7 +21,7 @@
 #include "source/opt/constants.h"
 #include "source/opt/fold.h"
 #include "source/opt/ir_context.h"
-#include "source/opt/make_unique.h"
+#include "source/util/make_unique.h"
 
 namespace spvtools {
 namespace opt {
