+++ source/opt/loop_descriptor.cpp
@@ -28,8 +28,8 @@
 #include "source/opt/ir_builder.h"
 #include "source/opt/ir_context.h"
 #include "source/opt/iterator.h"
-#include "source/opt/make_unique.h"
 #include "source/opt/tree_iterator.h"
+#include "source/util/make_unique.h"
 
 namespace spvtools {
 namespace opt {
