+++ source/opt/set_spec_constant_default_value_pass.cpp
@@ -22,9 +22,9 @@
 
 #include "source/opt/def_use_manager.h"
 #include "source/opt/ir_context.h"
-#include "source/opt/make_unique.h"
 #include "source/opt/type_manager.h"
 #include "source/opt/types.h"
+#include "source/util/make_unique.h"
 #include "source/util/parse_number.h"
 #include "spirv-tools/libspirv.h"
