+++ source/opt/ssa_rewrite_pass.cpp
@@ -46,8 +46,8 @@
 
 #include "source/opcode.h"
 #include "source/opt/cfg.h"
-#include "source/opt/make_unique.h"
 #include "source/opt/mem_pass.h"
+#include "source/util/make_unique.h"
 
 // Debug logging (0: Off, 1-N: Verbosity level).  Replace this with the
 // implementation done for
