+++ test/opt/iterator_test.cpp
@@ -18,7 +18,7 @@
 #include "gmock/gmock.h"
 
 #include "source/opt/iterator.h"
-#include "source/opt/make_unique.h"
+#include "source/util/make_unique.h"
 
 namespace spvtools {
 namespace opt {
